ajout de test

+++ Backend/controllers/books.js
@@ -2,6 +2,7 @@ const Book = require('../models/Books')
 const fs = require('fs')
 const sharp=require('sharp');
 const path=require('path');
+const mongoose=require('mongoose')
 
 exports.createBook=(req,res)=>{
     const bookObject = JSON.parse(req.body.book);
@@ -47,7 +48,11 @@ exports.getTopBooks=(req,res)=>{
 }
 
 exports.getBookbyId=(req,res)=>{
-       const bookId= req.params.id; 
+       const bookId= req.params.id;
+    
+       if (!mongoose.Types.ObjectId.isValid(bookId)) {
+        return res.status(400).json({ message: 'ID invalide' });
+       } 
         Book.findOne({_id:bookId})
           .then(book => {
               if (!book) {
@@ -61,15 +66,19 @@ exports.getBookbyId=(req,res)=>{
 
 exports.modifyBook=(req,res)=>{
     let bookObject;
+
+   
     Book.findOne({_id: req.params.id})
     .then(book=>{
 
     
-    if(!book) return res.status(404).json({message:'Livre non trouvé'});
+    if(!book) {
+      return res.status(404).json({message:'Livre non trouvé'});
+    }
   
 
     
-    if(book._userId !== req.auth.userId){
+    if(book.userId !== req.auth.userId){
         return res.status(401).json({message:'Non autorisé'});
     }
     
@@ -84,8 +93,11 @@ exports.modifyBook=(req,res)=>{
         .toFile(output)
         .then(()=>{const oldImage = book.imageUrl.split('/images/')[1];
             
-      
-       return fs.promises.unlink(`images/${oldImage}`)})
+          if (oldImage){
+          return fs.promises.unlink(`images/${oldImage}`).catch(() => null);
+          }
+          
+          })
        
 
         .then(()=>{
@@ -98,7 +110,9 @@ exports.modifyBook=(req,res)=>{
         delete bookObject._userId;
             return Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id });
           });
-      } else {
+      } 
+      
+    else {
         bookObject = { ...req.body };
         delete bookObject._userId;
         return Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id });
@@ -108,7 +122,9 @@ exports.modifyBook=(req,res)=>{
       res.status(200).json({ message: 'Livre modifié avec succès' });
     })
     .catch(error => {
+         if (!res.headersSent) {
       res.status(400).json({ message: 'Erreur lors de la modification', error });
+         }
     });
 };
 
@@ -118,24 +134,37 @@ exports.modifyBook=(req,res)=>{
 
 
 exports.deleteBook=(req,res)=>{
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(400).json({ message: 'ID invalide' });
+       } 
     Book.findOne({ _id: req.params.id })
     .then((book)=>{
+         if(!book){ 
+          return res.status(404).json({message:'Livre non trouvé'});
+        }
        if (book.userId != req.auth.userId)  
         {
             res.status(401).json({message:'Non-authorisé'})
         }
-        else
-        {
+
+         if(!book) return res.status(404).json({message:'Livre non trouvé'});
+        
+        
             const filename=book.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`,()=>{
             Book.deleteOne({_id: req.params.id})
                 .then(()=>res.status(200).json({message:'Objet supprimé'}))
-                .catch((error)=>res.status(401).json({error}))
+                
+                .catch((error)=>{
+                  if (!res.headersSent) {
+                  res.status(400).json({error})
+                  }
+                })
 
-            })
+            
           
 
-        }})
+    })})
 
     
     .catch(error => res.status(500).json({ error }));
@@ -144,20 +173,27 @@ exports.deleteBook=(req,res)=>{
 
 
 exports.rateBook=(req,res)=>{
+     if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        return res.status(400).json({ message: 'ID invalide' });
+       } 
     Book.findOne({_id: req.params.id})
     .then((book)=>{
         if (!book) {
             return res.status(404).json({ message: 'Livre non trouvé' });
         }
 
-        if(book.ratings.some((rating)=>rating.userId === req.auth.userId)){
-            return res.status(401).json({message:'Non-authorisé'})
-            
-
-        }
+        
         if (req.body.rating < 0 || req.body.rating > 5) {
         return res.status(400).json({ message: 'Note invalide (0 à 5 autorisés).' });
       }
+
+      
+
+        if(book.ratings.some((rating)=>rating.userId === req.auth.userId)){
+            return res.status(401).json({message:'Non-authorisé'})
+        }
+            
+
